requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
protobuf==3.20.1
pymongo==4.3.2
PyMySQL==1.0.2
python-dotenv==0.21.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
